[79]: refactor the Section component
