+++ common/uicomponent/src/main/java/com/github/tinkzhang/uicomponent/AmazonLinkSection.kt
@@ -2,18 +2,17 @@ package com.github.tinkzhang.uicomponent
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
-import androidx.compose.material.Card
-import androidx.compose.material.Divider
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Launch
-import androidx.compose.material3.*
+import androidx.compose.material3.FilledTonalButton
+import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.tooling.preview.Preview
-import androidx.compose.ui.unit.sp
 
 @Composable
 fun AmazonLinkSection(url: String?, mCustomTabClient: RkCustomTabClient?) {
@@ -45,44 +44,6 @@ fun AmazonLinkSection(url: String?, mCustomTabClient: RkCustomTabClient?) {
         }
     }
 
-}
-
-@Composable
-fun Section(
-    header: @Composable () -> Unit = {},
-    title: String? = null,
-    content: @Composable () -> Unit,
-) {
-    Card(Modifier.fillMaxWidth(), elevation = DpCardElevation) {
-        Column(
-            Modifier
-                .fillMaxWidth()
-        ) {
-            Surface(color = MaterialTheme.colorScheme.surfaceVariant) {
-                Column(
-                    Modifier
-                        .fillMaxWidth()
-                        .padding(DpCardInnerPadding)
-                ) {
-                    header()
-                    if (!title.isNullOrBlank()) {
-                        Text(
-                            text = title,
-                            style = MaterialTheme.typography.titleMedium.copy(
-                                fontSize = 18.sp,
-                                fontWeight = FontWeight.W500
-                            ),
-                        )
-                    }
-                    Divider(
-                        Modifier.padding(vertical = DpContentMediumPadding),
-                        thickness = DpDividerMedium
-                    )
-                    content()
-                }
-            }
-        }
-    }
 }
 
 @Preview
