+++ feature/reading/src/main/java/com/github/tinkzhang/reading/ui/uicomponents/ReadingVipNoteSection.kt
@@ -4,10 +4,8 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
-import androidx.compose.material.Divider
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.OutlinedButton
-import androidx.compose.material3.Surface
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
@@ -16,56 +14,45 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.github.tinkzhang.basic.model.PageFormat
 import com.github.tinkzhang.basic.model.ReadingRecord
+import com.github.tinkzhang.uicomponent.Section
 import com.google.firebase.Timestamp
 
 @Composable
 fun ReadingVipNoteSection(records: List<ReadingRecord>, pageFormat: PageFormat, pages: Int) {
-    Surface(tonalElevation = 4.dp) {
-        Column(
-            Modifier
-                .fillMaxWidth()
-                .padding(8.dp)
-        ) {
-            Text(
-                text = "Reading Notes",
-                style = MaterialTheme.typography.titleMedium
-            )
-            Divider(Modifier.padding(vertical = 4.dp), thickness = 2.dp)
-            when {
-                records.isEmpty() -> {
-                    Text(
-                        "\uD83D\uDDF3️  Empty reading note.",
-                        Modifier
-                            .align(Alignment.CenterHorizontally)
-                            .padding(16.dp),
-                        style = MaterialTheme.typography.titleSmall
-                    )
-                }
-                records.size < 4 -> {
-                    records.forEach {
-                        BookNote(it, pageFormat, pages)
-                    }
-                }
-                else -> {
-                    records.subList(0, 4).forEach {
-                        BookNote(it, pageFormat, pages)
-                    }
+    Section(title = "Reading Notes") {
+        when {
+            records.isEmpty() -> {
+                Text(
+                    "\uD83D\uDDF3️  Empty reading note.",
+                    Modifier
+                        .align(Alignment.CenterHorizontally)
+                        .padding(16.dp),
+                    style = MaterialTheme.typography.titleSmall
+                )
+            }
+            records.size < 4 -> {
+                records.forEach {
+                    BookNote(it, pageFormat, pages)
                 }
             }
-            if (records.isNotEmpty()) {
-                OutlinedButton(
-                    onClick = {},
-                    Modifier
-                        .fillMaxWidth(0.8f)
-                        .align(Alignment.CenterHorizontally)
-                        .padding(top = 8.dp)
-                ) {
-                    Text("Show All Notes")
+            else -> {
+                records.subList(0, 4).forEach {
+                    BookNote(it, pageFormat, pages)
                 }
             }
         }
+        if (records.isNotEmpty()) {
+            OutlinedButton(
+                onClick = {},
+                Modifier
+                    .fillMaxWidth(0.8f)
+                    .align(Alignment.CenterHorizontally)
+                    .padding(top = 8.dp)
+            ) {
+                Text("Show All Notes")
+            }
+        }
     }
-
 }
 
 @Composable
@@ -76,11 +63,16 @@ fun BookNote(record: ReadingRecord, pageFormat: PageFormat, pages: Int) {
                 .fillMaxWidth()
                 .padding(vertical = 8.dp)
         ) {
-            ReadingProgressCircleWithText(format = pageFormat, position = record.endPage, totalPages = pages)
+            ReadingProgressCircleWithText(
+                format = pageFormat,
+                position = record.endPage,
+                totalPages = pages
+            )
             Column(
                 Modifier
                     .fillMaxWidth()
-                    .padding(start = 8.dp)) {
+                    .padding(start = 8.dp)
+            ) {
                 Text(
                     record.note,
                     style = MaterialTheme.typography.bodyMedium
