+++ feature/reading/src/main/java/com/github/tinkzhang/reading/ui/uicomponents/ReadingVipProgressSection.kt
@@ -2,10 +2,7 @@ package com.github.tinkzhang.reading.ui.uicomponents
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.layout.*
-import androidx.compose.material.Divider
 import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Surface
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -16,6 +13,7 @@ import com.github.tinkzhang.basic.model.PageFormat
 import com.github.tinkzhang.basic.model.ReadingPlatform
 import com.github.tinkzhang.basic.model.ReadingRecord
 import com.github.tinkzhang.uicomponent.ReadingProgressText
+import com.github.tinkzhang.uicomponent.Section
 
 @Composable
 fun ReadingVipProgressSection(
@@ -24,44 +22,33 @@ fun ReadingVipProgressSection(
     totalPages: Int,
     platform: ReadingPlatform
 ) {
-    Surface(tonalElevation = 4.dp) {
-        Column(
-            Modifier
-                .fillMaxWidth()
-                .padding(8.dp)
+    Section(title = "Reading Progress") {
+        Box(
+            Modifier.align(Alignment.CenterHorizontally),
+            contentAlignment = Alignment.Center
         ) {
-            Text(
-                text = "Reading Progress",
-                style = MaterialTheme.typography.titleMedium
-            )
-            Divider(Modifier.padding(vertical = 4.dp), thickness = 2.dp)
-            Box(
-                Modifier.align(Alignment.CenterHorizontally),
-                contentAlignment = Alignment.Center
+            Column(
+                Modifier
+                    .fillMaxWidth()
+                    .padding(16.dp),
+                horizontalAlignment = Alignment.CenterHorizontally
             ) {
-                Column(
-                    Modifier
-                        .fillMaxWidth()
-                        .padding(16.dp),
-                    horizontalAlignment = Alignment.CenterHorizontally
-                ) {
-                    Image(
-                        painter = painterResource(id = platform.icon),
-                        contentDescription = platform.label
-                    )
-                    Spacer(modifier = Modifier.height(4.dp))
-                    ReadingProgressText(
-                        format = pageFormat,
-                        position = lastRecord?.endPage ?: 0,
-                        textStyle = MaterialTheme.typography.titleMedium,
-                    )
-                }
-
-                ReadingProgressCircle(
-                    Modifier.size(192.dp),
-                    progress = lastRecord?.endPage?.toFloat()?.div(totalPages.toFloat()) ?: 0.0f,
+                Image(
+                    painter = painterResource(id = platform.icon),
+                    contentDescription = platform.label
+                )
+                Spacer(modifier = Modifier.height(4.dp))
+                ReadingProgressText(
+                    format = pageFormat,
+                    position = lastRecord?.endPage ?: 0,
+                    textStyle = MaterialTheme.typography.titleMedium,
                 )
             }
+
+            ReadingProgressCircle(
+                Modifier.size(192.dp),
+                progress = lastRecord?.endPage?.toFloat()?.div(totalPages.toFloat()) ?: 0.0f,
+            )
         }
     }
 }
