+++ feature/wish/src/main/java/com/github/tinkzhang/wish/ui/components/VipSearchEngineSection.kt
@@ -3,11 +3,12 @@ package com.github.tinkzhang.wish.ui.components
 import android.net.Uri
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.shape.RoundedCornerShape
-import androidx.compose.material.Divider
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.Launch
 import androidx.compose.material.icons.filled.Search
-import androidx.compose.material3.*
+import androidx.compose.material3.Icon
+import androidx.compose.material3.OutlinedButton
+import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -16,6 +17,7 @@ import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.github.tinkzhang.firebaseRemoteConfig.SearchEngine
 import com.github.tinkzhang.uicomponent.RkCustomTabClient
+import com.github.tinkzhang.uicomponent.Section
 
 @Composable
 fun VipSearchEngineSection(
@@ -23,21 +25,10 @@ fun VipSearchEngineSection(
     searchEngines: List<SearchEngine>,
     client: RkCustomTabClient?
 ) {
-    Surface(tonalElevation = 4.dp) {
-        Column(
-            Modifier
-                .fillMaxWidth()
-                .padding(8.dp)
-        ) {
-            Text(
-                text = "Find the Book",
-                style = MaterialTheme.typography.titleMedium
-            )
-            Divider(Modifier.padding(vertical = 4.dp), thickness = 2.dp)
-            searchEngines.forEach {
-                Spacer(modifier = Modifier.height(8.dp))
-                SearchEngineButton(title = title, searchEngine = it, client)
-            }
+    Section(title = "Find the Book") {
+        searchEngines.forEach {
+            Spacer(modifier = Modifier.height(8.dp))
+            SearchEngineButton(title = title, searchEngine = it, client)
         }
     }
 }
